requirements.txt
# Python dependencies for Stock & ETF Visualization App
streamlit
requests
python-dotenv
